Form: extract Checkbox
Will be used in the replay UI.
